Add HuffmanTree support Update bcprov to 1.64 Update cglib to 3.3.0 Update log4j to 2.12.1 Update slf4j to 1.7.28 Update jackson to 2.10.0 Update snmp4j to 2.8.0 Update javamail to 1.6.4 Update jcifs to 2.1.3 Fix bug for pass FindBug analyze

+++ src/main/java/com/nervousync/utils/FileUtils.java
@@ -2541,7 +2541,7 @@ public final class FileUtils {
 				byteArrayOutputStream = new ByteArrayOutputStream(blockSize);
 				byteArrayOutputStream.write(readBuffer, 0, readLength);
 				SegmentationItem segmentationItem =
-						new SegmentationItem(index * blockSize, byteArrayOutputStream.toByteArray());
+						new SegmentationItem((long) index * blockSize, byteArrayOutputStream.toByteArray());
 				segmentationItemList.add(segmentationItem);
 				index++;
 			}
